float instead of int for time calculations

+++ src/gameplay_state.cpp
@@ -59,7 +59,7 @@ void GameplayState::render(Engine* engine) {
 void GameplayState::updateNextActor() {
     Actor* activeActor = actors->at(0);
 
-    int actionTime = activeActor->update(this);
+    float actionTime = activeActor->update(this);
     *activeActor->energy -= actionTime;
 
     actors->erase(actors->begin());
